Factored payment utilities out into multiple files
